+++ get_pid.m
@@ -2,17 +2,17 @@ function [ result, cumulative_err ] = get_pid( dt, curr_err, prev_err, cumulativ
 
     % constants
     kp = 0.6;
-    ki = 14;
+    ki = 14.0;
     kd = 0.005;
     
-    cumulative_err = cumulative_err + (curr_err*dt);
-    dif_err = (curr_err - prev_err)/dt;
+    cumulative_err = double(cumulative_err + (curr_err*dt));
+    dif_err = double((double(curr_err) - double(prev_err))/dt);
     
     a = kd * dif_err;
     b = ki * dif_err;
-    c = kp * curr_err;
+    c = kp * double(curr_err);
     
-    result = a + b + c;
+    result = double(a + b + c);
     
 
 end
